Remove %ldconfig_scriptlets (F28+ only)
%ldconfig_scriptlets does nothing on F28+.

The original idea was to keep is present not to unnecessarily diverge master
and f27, but since that have happened anyway, I'm removing it for spec
simplicity. Not adding a new changelog entry (it's not user facing) nor bumping
the release (it doesn't need to be rebuilt, it changes nothing). Altering the
changelog entry that was added when %ldconfig_scriptlets was added.