POC: keep only cache for large, autogenerated files
This saves about 2 MiB from the package. When all rpms get extracted,
the comparison is as follows on an ext4 filesystem:

With this patch:
$ du -s usr
195828 usr/

Without this patch:
$ du -s usr
198224 usr/

Which is a difference of 2.396 MiB.

To reduce the filesystem footprint of the Python installation, some
files are now being removed and only one level of their pycache
(non-optimized) is being kept.

These particular files were chosen for their size and the fact that they
are autogenerated, thus hard to read.