Use %autosetup with git am
Patch 189 is still applied conditionally, so we need to use %autosetup with -N,
to prevent automatic patch application.

Later, the patches are applied automatically up to 188 (with -M for maximum),
patch 189 is applied conditionally,
and finally patches from 190 up are applied (with -m for minimum).

Once patch 189 is reworked to be applicable at all times (or removed), this can be dropped.

While not necessary shorter, this no longer requires adding/removing %patch XYZ
when adding/removing patches.

While at it, we also use git am to apply the patches
since they are all generated from git anyway.
This makes it easier if we ever patch binary files.