Stop Recommending setuptools and move the pip Recommendation to python3-devel
With modern pip, setuptools is not required on runtime.
Python packages use PEP 517 instead.

Hence, we no longer Recommend setuptools and pip, but pip only.

Users who are impacted may see one of:

    ModuleNotFoundError: No module named 'setuptools'
    ModuleNotFoundError: No module named 'pkg_resoureces'

They can easily recover by installing python3-setuptools or python3-pkg_resources.

Related but not explicitly part of:
https://fedoraproject.org/wiki/Changes/Reduce_dependencies_on_python3-setuptools

Together with this change, we move the pip Recommendation to the python3-devel package.
"Regular users" of Fedora are less likely to need pip and more likely to brick
their systems with it. If they need pip, they can dnf install pip.

Pip (incl. setuptools) still works in environments created with venv or virtualenv.

Not bumping the release because this only affects the package when we
upgrade the main Python in Fedora to 3.10.